[REFACTOR] 리뷰 반영 - createIndexIfNotExists 분리
