+++ user-api/src/main/java/com/biengual/userapi/content/domain/ContentSearchRepositoryImpl.java
@@ -9,8 +9,6 @@ import java.util.List;
 
 import org.opensearch.client.opensearch.OpenSearchClient;
 import org.opensearch.client.opensearch._types.FieldValue;
-import org.opensearch.client.opensearch._types.Time;
-import org.opensearch.client.opensearch._types.mapping.TypeMapping;
 import org.opensearch.client.opensearch._types.query_dsl.BoolQuery;
 import org.opensearch.client.opensearch._types.query_dsl.FuzzyQuery;
 import org.opensearch.client.opensearch._types.query_dsl.NestedQuery;
@@ -23,79 +21,17 @@ import org.opensearch.client.opensearch.core.IndexResponse;
 import org.opensearch.client.opensearch.core.SearchRequest;
 import org.opensearch.client.opensearch.core.SearchResponse;
 import org.opensearch.client.opensearch.core.search.Hit;
-import org.opensearch.client.opensearch.indices.CreateIndexRequest;
-import org.opensearch.client.opensearch.indices.IndexSettings;
 import org.springframework.stereotype.Repository;
 
 import com.biengual.core.domain.document.content.ContentSearchDocument;
 import com.biengual.core.response.error.exception.CommonException;
 
-import jakarta.annotation.PostConstruct;
 import lombok.RequiredArgsConstructor;
 
 @Repository
 @RequiredArgsConstructor
 public class ContentSearchRepositoryImpl implements ContentSearchRepository {
     private final OpenSearchClient openSearchClient;
-
-    /**
-     * 인덱스가 없을 경우 인덱스 초기화
-     */
-    @PostConstruct
-    public void init() {
-        this.createIndexIfNotExists(openSearchClient);
-    }
-
-    private void createIndexIfNotExists(OpenSearchClient client) {
-        try {
-            boolean exists = client.indices().exists(b -> b.index(CONTENT_SEARCH_INDEX_NAME)).value();
-            if (!exists) {
-                // Define index settings
-                // TODO: 성능 부족하면 샤드, 레플리카 추가해야 하는지 고려
-                IndexSettings settings = new IndexSettings.Builder()
-                    .numberOfShards("1") // Set number of shards
-                    .numberOfReplicas("1") // Set number of replicas
-                    .refreshInterval(new Time.Builder().time("30s").build())
-                    .analysis(a -> a
-                        .analyzer("nori_analyzer", na -> na
-                            .custom(c -> c
-                                .tokenizer("nori_tokenizer")
-                                .filter("nori_part_of_speech", "lowercase")
-                            )
-                        )
-                        .analyzer("custom_analyzer", ca -> ca
-                            .custom(c -> c
-                                .tokenizer("standard")
-                                .filter("lowercase")
-                            )
-                        )
-                    )
-                    .build();
-
-                // Define index mappings
-                TypeMapping mappings = new TypeMapping.Builder()
-                    .properties("id", p -> p.keyword(k -> k.index(false)))
-                    .properties("title", p -> p.text(t -> t.analyzer("custom_analyzer")))
-                    .properties("categoryName", p -> p.text(t -> t.analyzer("custom_analyzer")))
-                    .properties("scripts", p -> p.nested(n -> n
-                        .properties("enScript", sp -> sp.text(t -> t.analyzer("custom_analyzer")))
-                        .properties("koScript", sp -> sp.text(t -> t.analyzer("nori_analyzer")))
-                    ))
-                    .build();
-
-                // Create the index with settings and mappings
-                CreateIndexRequest createIndexRequest = new CreateIndexRequest.Builder()
-                    .index(CONTENT_SEARCH_INDEX_NAME)
-                    .settings(settings)
-                    .mappings(mappings)
-                    .build();
-
-                client.indices().create(createIndexRequest);
-            }
-        } catch (Exception e) {
-            throw new CommonException(SEARCH_CONTENT_SAVE_FAILED);
-        }
-    }
 
     /**
      * 키워드로 콘텐츠를 검색합니다.
